third_party/WebKit/LayoutTests/platform/android/fast/images/image-zoom-to-500-expected.txt

This tests that page zoom and image auto-sizing interact well together. This test requires testRunner to run. To test manually, open this image in a browser window, resize the window to 600px tall and zoom in 10 times. The image should get bigger at each step proportionately.

Size at zoom level0 : 320x239.984375
Size at zoom level1 : 319.984375x239.984375
Size at zoom level2 : 319.984375x239.984375
Size at zoom level3 : 319.984375x239.984375
Size at zoom level4 : 319.984375x239.984375
Size at zoom level5 : 319.984375x239.984375
Size at zoom level6 : 319.984375x239.984375
Size at zoom level7 : 319.984375x239.984375
Size at zoom level8 : 320x240
Size at zoom level9 : 320x240
